YETUS-452. Remove python2; rewrite python bits for python3
