+++ .circleci/config.yml
@@ -40,12 +40,11 @@ jobs:
              --plugins=all
              --java-home=/usr/lib/jvm/java-8-openjdk-amd64
              --patch-dir=/tmp/yetus-out
-             --pylint=pylint2
              --html-report-file=/tmp/yetus-out/report.html
              --console-report-file=/tmp/yetus-out/console.txt
              --brief-report-file=/tmp/yetus-out/brief.txt
              --bugcomments=briefreport,htmlout,junit
-             --tests-filter=checkstyle,javadoc,rubocop,test4tests
+             --tests-filter=checkstyle,test4tests
              --junit-report-xml=/tmp/yetus-out/junit-report.xml
 
       - store_test_results:
