+++ .gitlab-ci.yml
@@ -23,13 +23,12 @@ buretoolbox-job:
       precommit/src/main/shell/test-patch.sh
       --patch-dir=/tmp/yetus-out
       --plugins=all
-      --pylint=pylint2
       --java-home=/usr/lib/jvm/java-8-openjdk-amd64
       --html-report-file=/tmp/yetus-out/report.html
       --console-report-file=/tmp/yetus-out/console.txt
       --brief-report-file=/tmp/yetus-out/brief.txt
       --bugcomments=briefreport,htmlout,gitlab,junit
-      --tests-filter=checkstyle,javadoc,rubocop,test4tests
+      --tests-filter=checkstyle,test4tests
       --junit-report-xml=/tmp/yetus-out/junit-report.xml
 
   artifacts:
